@@ -1,12 +1,9 @@
 from __future__ import absolute_import
-import calendar
 import inspect
 import re
 from datetime import date, datetime, time
 from decimal import Decimal
 from functools import wraps
-
-from six import integer_types, string_types
 
 from .fields import (FIELD_DATETIME, FIELD_NO_INPUT, FIELD_NUMERIC,
                      FIELD_SELECT, FIELD_SELECT_MULTIPLE, FIELD_TEXT,
@@ -70,7 +67,7 @@ class StringType(BaseType):
 
     def _assert_valid_value_and_cast(self, value):
         value = value or ""
-        if not isinstance(value, string_types):
+        if not isinstance(value, str):
             raise AssertionError("{0} is not a valid string type.".
                                  format(value))
         return value
@@ -115,7 +112,7 @@ class NumericType(BaseType):
         if isinstance(value, float):
             # In python 2.6, casting float to Decimal doesn't work
             return float_to_decimal(value)
-        if isinstance(value, integer_types):
+        if isinstance(value, int):
             return Decimal(value)
         if isinstance(value, Decimal):
             return value
@@ -175,8 +172,8 @@ class SelectType(BaseType):
 
     @staticmethod
     def _case_insensitive_equal_to(value_from_list, other_value):
-        if isinstance(value_from_list, string_types) and \
-                isinstance(other_value, string_types):
+        if isinstance(value_from_list, str) and \
+                isinstance(other_value, str):
             return value_from_list.lower() == other_value.lower()
         else:
             return value_from_list == other_value
