@@ -1,5 +1,5 @@
 from __future__ import absolute_import
-from datetime import time, datetime, timedelta
+from datetime import time, datetime
 
 from business_rules.operators import TimeType
 from .. import TestCase
